chore: inline ExclusiveRelationship mixin into Did
The mixin generalized mutual-exclusion handling for N pairs via
mapping.each, but has exactly one caller with exactly one pair
(voice_in_trunk <-> voice_in_trunk_group on Did). Inline the same
logic directly on Did and delete the mixin.
