+++ lib/didww/client.rb
@@ -131,7 +131,6 @@ module DIDWW
         require 'didww/resource/country'
         require 'didww/resource/did_group_type'
         require 'didww/resource/did_group'
-        require 'didww/resource/exclusive_relationship'
         require 'didww/resource/did'
         require 'didww/resource/order'
         require 'didww/resource/pop'
